tickets/src/routes/update.ts
import express, { Request, Response } from "express";
import mongoose from "mongoose";
import { body } from "express-validator";
import {
  validateRequest,
  requireAuth,
  NotFoundError,
  NotAuthorizedError,
} from "@akilaticketstest/common";
import Ticket from "../models/ticket";

const router = express.Router();

router.put(
  "/api/tickets/:id",
  requireAuth,
  [
    body("title").not().isEmpty().withMessage("Title is required"),
    body("price")
      .isFloat({ gt: 0 })
      .withMessage("Price must be provided and must be greater than 0"),
  ],
  validateRequest,
  async (req: Request, res: Response) => {
    const ticket = await Ticket.findById(req.params.id);

    if (!ticket) throw new NotFoundError();

    if (ticket.userId !== req.currentUser!.id) throw new NotAuthorizedError();

    // const updatedTicket = await Ticket.findByIdAndUpdate(
    //   { _id: req.params.id },
    //   {
    //     title: req.body.title,
    //     price: req.body.price,
    //   },
    //   { new: true }
    // );
    // res.send(updatedTicket);

    ticket.set({
      title: req.body.title,
      price: req.body.price,
    });

    await ticket.save();

    res.send(ticket);
  }
);

export { router as updateTicketRouter };
